Fixed "enhanced" typo, hyphenated "front-end" for consistency, added better sentence structure
So...I couldn't find the August 8th post in "sodevious-blog/_posts", but luckily enough, there were typos in the previous post for me to fix/change/try to suggest better sentence structure.

For the last paragraph, the best sentence I was able to come up with was: 

"Sometimes, there are problems where I'm scratching my head trying to understand how to fix them. The only way to make something better is to understand how and why it's bad in the first place."

The word "understand" is used twice, and while it makes sense perhaps using a synonym might be better.